created the movies app and moved movies.html and search.html
The Movies app now contains code related to the movies page and the search page. The next change that will be made here will be moving and renaming landing_page/index.css and index.js because it seems more relevant to the movies app.
