--- landing_page/static/landing_page/js/search.js
@@ -1,4 +1,0 @@
-function loadDetailsPage(movie_id) {
-    // movie = JSON.parse(jsonMovie);
-    window.location.href = window.location.origin + `/movie/${movie_id}`;
-}
