+++ landing_page/templates/landing_page/base.html
@@ -11,11 +11,9 @@
       crossorigin="anonymous"
     ></script>
     {% block stylesheets %} {% endblock %} {% block scripts %} {% endblock %}
-    <!-- <script src="{% static 'landing_page/js/index.js' %}" defer></script>
-    <link rel="stylesheet" href="{% static 'landing_page/css/index.css' %}" />
-    <link rel="stylesheet" href="{% static 'landing_page/css/about.css' %}" /> -->
     <link rel="stylesheet" href="{% static 'landing_page/css/base.css' %}" />
-    <script src="{% static 'landing_page/js/search.js' %}" defer></script>
+    <script src="{% static 'movies/js/search.js' %}" defer></script>
+    <script src="{% static 'landing_page/js/index.js' %}" defer></script>
 
     <title>ReeldIn</title>
   </head>
