+++ landing_page/templates/landing_page/index.html
@@ -1,7 +1,5 @@
 {% extends 'landing_page/base.html' %} {% load static %} {% block stylesheets %}
 <link rel="stylesheet" href="{% static 'landing_page/css/index.css' %}" />
-{% endblock %} {% block scripts %}
-<script src="{% static 'landing_page/js/index.js' %}" defer></script>
 {% endblock %} {% block content %}
 <div id="message-div">
   <h1>Welcome to ReeldIn</h1>
