+++ movies/static/movies/js/search.js
@@ -1,0 +1,4 @@
+function loadDetailsPage(movie_id) {
+  // movie = JSON.parse(jsonMovie);
+  window.location.href = window.location.origin + `/movie/${movie_id}`;
+}
