+++ movies/templates/movies/movie.html
@@ -1,8 +1,8 @@
 {% extends 'landing_page/base.html' %} {% load static %} {% block stylesheets %}
 <link rel="stylesheet" href="{% static 'accounts/css/modal.css' %}" />
-<link rel="stylesheet" href="{% static 'landing_page/css/movie.css' %}" />
+<link rel="stylesheet" href="{% static 'movies/css/movie.css' %}" />
 {% endblock %} {% block scripts %}
-<script src="{% static 'landing_page/js/movie.js' %}" defer></script>
+<script src="{% static 'movies/js/movie.js' %}" defer></script>
 {% endblock %} {% block content %}
 <div class="content-div">
   <div id="movie-name">
