+++ movies/templates/movies/search.html
@@ -1,5 +1,5 @@
 {% extends 'landing_page/base.html' %} {% load static %} {% block stylesheets %}
-<link rel="stylesheet" href="{% static 'landing_page/css/search.css' %}" />
+<link rel="stylesheet" href="{% static 'movies/css/search.css' %}" />
 {% endblock %} {% block scripts %}
 <script>
   var WATCHMODE_API_KEY = "{{ WATCHMODE_API_KEY }}";
